Move KVStore to separate file
