+++ src/app/app.cpp
@@ -9,189 +9,13 @@
 #include "ccf/json_handler.h"
 #include "etcd.pb.h"
 #include "grpc.h" // TODO(#25): use grpc from ccf
-#include "kv/untyped_map.h" // TODO(#22): private header
-
-#include <nlohmann/json.hpp>
+#include "kvstore.h"
 
 #define FMT_HEADER_ONLY
 #include <fmt/format.h>
 
 namespace app
 {
-  using json = nlohmann::json;
-
-  struct Value
-  {
-    // the actual value that the client wants written.
-    std::string value;
-    // the revision that this entry was created (since the last delete).
-    uint64_t create_revision;
-    // the latest modification of this entry (0 in the serialised field).
-    uint64_t mod_revision;
-    // the version of this key, reset on delete and incremented every update.
-    uint64_t version;
-
-    Value(std::string v)
-    {
-      value = v;
-      create_revision = 0;
-      mod_revision = 0;
-      version = 1;
-    }
-
-    Value()
-    {
-      Value("");
-    }
-  };
-  DECLARE_JSON_TYPE(Value);
-  DECLARE_JSON_REQUIRED_FIELDS(Value, value, create_revision, version);
-
-  static constexpr auto RECORDS = "public:records";
-
-  /// @brief KVStore is a wrapper around a CCF map that handles serialisation as
-  /// well as ensuring values have proper revisions when returned.
-  class KVStore
-  {
-  public:
-    using K = std::string;
-    using V = Value;
-    using KSerialiser = kv::serialisers::BlitSerialiser<K>;
-    using VSerialiser = kv::serialisers::JsonSerialiser<V>;
-
-    // Use untyped map so we can access the range API.
-    using MT = kv::untyped::Map;
-
-    /// @brief Constructs a KVStore
-    /// @param ctx
-    KVStore(kv::Tx& tx)
-    {
-      inner_map = tx.template rw<KVStore::MT>(RECORDS);
-    }
-    /// @brief Constructs a KVStore
-    /// @param ctx
-    KVStore(kv::ReadOnlyTx& tx)
-    {
-      inner_map = tx.template ro<KVStore::MT>(RECORDS);
-    }
-
-    /// @brief get retrieves the value stored for the given key. It hydrates the
-    /// value with up-to-date information as values may not store all
-    /// information about revisions.
-    /// @param key the key to get.
-    /// @return The value, if present.
-    std::optional<V> get(const K& key)
-    {
-      // get the value out and deserialise it
-      auto res = inner_map->get(KSerialiser::to_serialised(key));
-      if (!res.has_value())
-      {
-        return std::nullopt;
-      }
-      auto val = VSerialiser::from_serialised(res.value());
-
-      hydrate_value(key, val);
-
-      return val;
-    }
-
-    void range(
-      const std::function<void(K&, V&)>& fn, const K& from, const K& to)
-    {
-      inner_map->range(
-        [&](auto& key, auto& value) {
-          auto k = KSerialiser::from_serialised(key);
-          auto v = VSerialiser::from_serialised(value);
-          hydrate_value(k, v);
-          fn(k, v);
-        },
-        KSerialiser::to_serialised(from),
-        KSerialiser::to_serialised(to));
-    }
-
-    /// @brief Associate a value with a key in the store, replacing existing
-    /// entries for that key.
-    ///
-    /// When an entry doesn't exist already this simply writes the data in.
-    ///
-    /// When an entry does exist already this gets the old value, and uses the
-    /// data to build the new version and, if not set, a create revision.
-    /// @param key where to insert
-    /// @param value data to insert
-    /// @return the old value associated with the key, if present.
-    std::optional<V> put(K key, V value)
-    {
-      const auto old = this->get(key);
-      if (old.has_value())
-      {
-        const auto old_val = old.value();
-        if (old_val.create_revision == 0)
-        {
-          // first put after creation of this key so set the revision
-          auto version_opt = inner_map->get_version_of_previous_write(
-            KSerialiser::to_serialised(key));
-          if (version_opt.has_value())
-          {
-            // can set the creation revision
-            value.create_revision = version_opt.value();
-          }
-        }
-        else
-        {
-          // otherwise just copy it to the new value so that we don't lose it
-          value.create_revision = old_val.create_revision;
-        }
-
-        value.version = old_val.version + 1;
-      }
-
-      inner_map->put(
-        KSerialiser::to_serialised(key), VSerialiser::to_serialised(value));
-
-      return old;
-    }
-
-    /// @brief remove data associated with the key from the store.
-    /// @param key
-    /// @return the old value, if present in the store.
-    std::optional<V> remove(const K& key)
-    {
-      auto k = KSerialiser::to_serialised(key);
-      auto old = inner_map->get(k);
-      inner_map->remove(k);
-      if (old.has_value())
-      {
-        return VSerialiser::from_serialised(old.value());
-      }
-      else
-      {
-        return std::nullopt;
-      }
-    }
-
-  private:
-    MT::Handle* inner_map;
-
-    void hydrate_value(const K& key, V& value)
-    {
-      // the version of the write to this key is our revision
-      auto version_opt = inner_map->get_version_of_previous_write(
-        KSerialiser::to_serialised(key));
-      // if there is no version (somehow) then just default it
-      // this shouldn't be nullopt though.
-      uint64_t revision = version_opt.value_or(0);
-
-      // if this was the first insert then we need to set the creation revision.
-      if (value.create_revision == 0)
-      {
-        value.create_revision = revision;
-      }
-
-      // and always set the mod_revision
-      value.mod_revision = revision;
-    }
-  };
-
   class AppHandlers : public ccf::UserEndpointRegistry
   {
   public:
@@ -301,7 +125,7 @@ namespace app
             payload.max_create_revision()));
       }
 
-      auto records_map = KVStore(ctx.tx);
+      auto records_map = store::KVStore(ctx.tx);
       auto& key = payload.key();
       auto& range_end = payload.range_end();
       if (range_end.empty())
@@ -391,8 +215,8 @@ namespace app
           fmt::format("ignore lease not yet supported"));
       }
 
-      auto records_map = KVStore(ctx.tx);
-      records_map.put(payload.key(), Value(payload.value()));
+      auto records_map = store::KVStore(ctx.tx);
+      records_map.put(payload.key(), store::Value(payload.value()));
       ctx.rpc_ctx->set_response_status(HTTP_STATUS_OK);
 
       return ccf::grpc::make_success(put_response);
@@ -412,7 +236,7 @@ namespace app
         payload.prev_kv());
       etcdserverpb::DeleteRangeResponse delete_range_response;
 
-      auto records_map = KVStore(ctx.tx);
+      auto records_map = store::KVStore(ctx.tx);
       auto& key = payload.key();
 
       if (payload.range_end().empty())
